requirements.txt
streamlit
pandas             # For data manipulation
matplotlib         # For plotting
google-cloud-bigquery